Add SendIssueDropdowns component
