+++ rdmo/projects/assets/js/project/components/areas/dashboard/SendIssueModal.js
@@ -5,6 +5,8 @@ import { useSelector } from 'react-redux'
 import { Modal } from 'rdmo/core/assets/js/_bs53/components'
 import { Input, Textarea } from 'rdmo/core/assets/js/components/forms'
 
+import SendIssueDropdowns from './SendIssueDropdowns'
+
 const SendIssueModal = ({
   issue,
   onClose
@@ -13,15 +15,6 @@ const SendIssueModal = ({
   const currentUser = useSelector((state) => state.user.currentUser) ?? {}
   const settings = useSelector(state => state.settings)
   const sites = useSelector(state => state.sites) ?? {}
-  const answers = useSelector((state) => state.project.project.answers) ?? {}
-  const views = useSelector((state) => state.project.project.views) ?? []
-  const snapshots = useSelector((state) => state.project.project.snapshots) ?? []
-  const snapshotOptions = [
-    { id: 'current', title: gettext('Current') },
-    ...snapshots
-  ]
-  const files = []
-  const formats = settings.export_formats ?? []
   const currentSite = Object.values(sites).find((site) => site.id === project.site)
 
   const initialMessage = [
@@ -127,177 +120,12 @@ const SendIssueModal = ({
         {
           isConfigured && (
             <>
-              <div className="d-flex justify-content-end gap-3 mb-3">
-                <div className="dropdown">
-                  <button
-                    type="button"
-                    className="link text-nowrap"
-                    data-bs-toggle="dropdown"
-                    data-bs-popper-config='{"strategy":"fixed"}'
-                    aria-expanded="false"
-                    title={gettext('Attachments')}
-                  >
-                    {gettext('Attachments')} <i className="bi bi-caret-down-fill ms-1" />
-                  </button>
-
-                  <div className="dropdown-menu p-3" style={{ minWidth: '280px' }}>
-                    {
-                      answers.html && (
-                        <div className="mb-3">
-                          <div className="fw-bold mb-2">
-                            {gettext('Answers')}
-                          </div>
-
-                          <div className="form-check">
-                            <input
-                              id="id_attachments_answers"
-                              name="attachments_answers"
-                              type="checkbox"
-                              className="form-check-input"
-                              value="project_answers"
-                              checked={formData.attachments_answers.includes('project_answers')}
-                              onChange={
-                                (event) => handleCheckboxChange(
-                                  'attachments_answers',
-                                  'project_answers',
-                                  event.target.checked
-                                )
-                              }
-                            />
-                            <label
-                              className="form-check-label fw-normal"
-                              htmlFor="id_attachments_answers"
-                            >
-                              {/* TODO: use a template for that? */}
-                              {gettext('Attach the output of "View answers".')}
-                            </label>
-                          </div>
-                        </div>
-                      )
-                    }
-                    {
-                      views.length > 0 && (
-                        <div className="mb-3">
-                          <div className="fw-bold mb-2">{gettext('Views')}</div>
-                          {
-                            views.map((view) => (
-                              <div className="form-check" key={view.id}>
-                                <input
-                                  id={`id_attachments_views_${view.id}`}
-                                  name="attachments_views"
-                                  type="checkbox"
-                                  className="form-check-input"
-                                  value={view.id}
-                                  checked={formData.attachments_views.includes(view.id)}
-                                  onChange={
-                                    (event) => handleCheckboxChange('attachments_views', view.id, event.target.checked)
-                                  }
-                                />
-                                <label
-                                  className="form-check-label fw-normal"
-                                  htmlFor={`id_attachments_views_${view.id}`}>
-                                  {view.title}
-                                </label>
-                              </div>
-                            ))
-                          }
-                        </div>
-                      )
-                    }
-
-                    {
-                      files.length > 0 && (
-                        <div className="mb-3">
-                          <div className="fw-bold mb-2">{gettext('Files')}</div>
-                          {
-                            files.map((file) => (
-                              <div className="form-check" key={file.id}>
-                                <input
-                                  id={`id_attachments_files_${file.id}`}
-                                  name="attachments_files"
-                                  type="checkbox"
-                                  className="form-check-input"
-                                  value={file.id}
-                                  checked={formData.attachments_files.includes(file.id)}
-                                  onChange={
-                                    (event) => handleCheckboxChange('attachments_files', file.id, event.target.checked)
-                                  }
-                                />
-                                <label
-                                  className="form-check-label fw-normal"
-                                  htmlFor={`id_attachments_files_${file.id}`}>
-                                  {file.file_name}
-                                </label>
-                              </div>
-                            ))
-                          }
-                        </div>
-                      )
-                    }
-
-                    <div>
-                      <div className="fw-bold mb-2">{gettext('Snapshot')}</div>
-                      {
-                        snapshotOptions.map((snapshot) => (
-                          <div className="form-check" key={snapshot.id}>
-                            <input
-                              id={`id_attachments_snapshot_${snapshot.id}`}
-                              name="attachments_snapshot"
-                              type="radio"
-                              className="form-check-input"
-                              value={snapshot.id}
-                              checked={formData.attachments_snapshot === snapshot.id}
-                              onChange={() => setField('attachments_snapshot', snapshot.id)}
-                            />
-                            <label
-                              className="form-check-label fw-normal"
-                              htmlFor={`id_attachments_snapshot_${snapshot.id}`}
-                            >
-                              {snapshot.title}
-                            </label>
-                          </div>
-                        ))
-                      }
-                    </div>
-                  </div>
-                </div>
-                <div className="dropdown">
-                  <button
-                    type="button"
-                    className="link text-nowrap"
-                    data-bs-toggle="dropdown"
-                    data-bs-popper-config='{"strategy":"fixed"}'
-                    aria-expanded="false"
-                    title={gettext('Format')}
-                  >
-                    {gettext('Format')} <i className="bi bi-caret-down-fill ms-1" />
-                  </button>
-
-                  <div className="dropdown-menu p-3" style={{ minWidth: '220px' }}>
-                    {
-                      formats.map(([format, label]) => (
-                        <div className="form-check" key={format}>
-                          <input
-                            id={`id_attachments_format_${format}`}
-                            name="attachments_format"
-                            type="radio"
-                            className="form-check-input"
-                            value={format}
-                            checked={formData.attachments_format === format}
-                            onChange={() => setField('attachments_format', format)}
-                          />
-                          <label
-                            className="form-check-label fw-normal"
-                            htmlFor={`id_attachments_format_${format}`}
-                          >
-                            {label}
-                          </label>
-                        </div>
-                      ))
-                    }
-                  </div>
-                </div>
-              </div>
+              <SendIssueDropdowns
+                formData={formData}
+                setField={setField}
+                onCheckboxChange={handleCheckboxChange}
+                formats={settings.export_formats ?? []}
+              />
               {/* TODO: use a template for that? */}
               <p>{gettext('Sending a task will set the status to "in progress".')}</p>
               <Input
